Rework into new layout as well as proper navbar utilization
This reworks the folder structure as well as the sidebar to work
alongside the navbar.

This makes mobile navigation easier.

The config file changes should allow everyone the old niceities of not
specifying the path for every bloody file while having an actual folder
structure that makes sense.

This is already a feature in vuepress v2 but oh well, here's my
reimplementation of that at least.

The new thing to edit for contributors will be arrays.js.